Do the arithmetic in base ten
`0.1 + 0.2 - 0.3` was 5.5511151231258e-17, and no amount of display
rounding was ever going to fix that: a tenth has no exact binary
representation, so the error is in the addition, not in the printing.
Apple's calculator does not have the problem because its basic
arithmetic is decimal. Now neither does this one.

`core/src/engine/decimal.rs` is the new number: an i128 mantissa of at
most eighteen significant digits and an i32 exponent, `m × 10^e`.
Addition, subtraction, multiplication, division, percent, modulo,
whole-number powers and factorials up to 22! are carried out in it, so
a number a person can write down is a number the calculator holds
exactly. Division rounds where it must; everything else is exact.

Eighteen digits carried against fifteen displayed makes three guard
digits, which is what a decimal calculator has always used the extra
digits for: the rounding of `1÷3` lands three places past anything on
screen, so `1÷3 = ×3 =` still comes back as 1.

The trigonometry, the logarithms, the roots and the fractional powers
have no decimal algorithm worth writing, so they hand their argument
to f64 and read the answer back as the shortest decimal that
identifies the double — `√0.01` re-enters as `0.1`, not as
`0.1000000000000000055`, so the arithmetic after it is exact again.
The range stays the double one: values past ~1e308 report Overflow and
non-zero values below 1e-308 report Underflow, exactly where they did.

Three test expectations moved, all of them pinned to binary artifacts:
a memory register whose two subnormal operands now differ by exactly
1e-313 (below the range, so it says Underflow like the evaluator does
for the same expression), and a `mod 3` of a thirteen-digit product
whose residue was 2⁻¹⁰ and is now a decimal tail. `1e15 + 1` still
shows all sixteen of its digits.

Also drops "Automatic scientific mode in landscape window" from the
README's feature list. There was never any code for it.

+++ src/ui/buttons.rs
@@ -17,7 +17,7 @@
 
 use crate::config::{Config, Mode};
 use crate::engine::item::{BinOp, BinaryFunc, ConstKind, InputItem, UnaryFunc};
-use crate::engine::{AngleMode, Engine};
+use crate::engine::{AngleMode, Decimal, Engine};
 use crate::rng::rand_value;
 
 /// Maximum number of significant digits the user is allowed to type
@@ -156,7 +156,7 @@ pub struct UiState {
     /// Numeric value behind `last_result`, used by `M+` / `M-` when the
     /// user stores the result of a just-evaluated expression without
     /// clearing the buffer first.
-    pub last_result_value: Option<f64>,
+    pub last_result_value: Option<Decimal>,
     /// Previous evaluated expression (and any error message from the
     /// most recent `=`). Shown as a caption above the main display; the
     /// main display itself renders the buffer, which now holds the
@@ -1522,11 +1522,12 @@ pub fn insert_number_string(engine: &mut Engine, s: &str) -> (usize, usize) {
 ///
 /// This is what keeps a result usable as an operand. `1÷3=` shows
 /// 0.333333333333333 either way, but multiplying those fifteen digits
-/// by three gives 0.999999999999999, while multiplying the value they
-/// stand for gives the 1 the user is expecting. The exactness lasts
-/// only as long as the digits are untouched — edit them and the buffer
-/// drops the annotation, because then they really are the number.
-pub fn insert_exact_value(engine: &mut Engine, shown: &str, value: f64) {
+/// by three gives 0.999999999999999, while multiplying the eighteen
+/// the division was computed to gives the 1 the user is expecting.
+/// The exactness lasts only as long as the digits are untouched — edit
+/// them and the buffer drops the annotation, because then they really
+/// are the number.
+pub fn insert_exact_value(engine: &mut Engine, shown: &str, value: Decimal) {
     let (start, end) = insert_number_string(engine, shown);
     engine.input.mark_exact(start, end, value);
 }
